code/teethanalyzer-backend/requirements.txt
fastapi
uvicorn
torch
torchvision
Pillow
python-multipart
timm
pydantic
python-dotenv
google-generativeai